fastapi/pyproject.toml
[project]
name = "projx-fastapi"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
requires-python = ">3.10,<3.13"
dependencies = [
    "fastapi>=0.121.1,<0.122.0",
    "starlette>=0.49.1,<0.50.0",
    "uvicorn>=0.35.0,<0.36.0",
    "sqlalchemy>=2.0.43,<3.0.0",
    "alembic>=1.16.5,<2.0.0",
    "asyncpg>=0.30.0,<0.31.0",
    "greenlet>=3.2.4,<4.0.0",
    "python-dotenv>=1.1.1,<2.0.0",
    "loguru>=0.7.3,<0.8.0",
    "pydantic>=2.11.7,<3.0.0",
    "aiomysql>=0.2.0,<0.4.0",
    "pyjwt>=2.10.1,<3.0.0",
    "cryptography>=46.0.5,<47.0.0",
]

[tool.uv]
package = false
default-groups = ["dev", "test"]

[dependency-groups]
test = [
    "pytest>=9.0.3,<10.0.0",
    "pytest-asyncio>=1.3.0,<2.0.0",
    "httpx>=0.27.0,<0.28.0",
    "pytest-cov>=4.1.0,<5.0.0",
    "faker>=24.0.0,<25.0.0",
]
dev = [
    "import-linter>=2.0",
    "ruff>=0.11.0,<1.0.0",
    "mypy>=1.17.1,<2.0.0",
    "pip-audit>=2.9.0,<3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --strict-markers --cov=src --cov-report=term-missing --cov-fail-under=80"
markers = [
    "asyncio: mark test as async",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "src/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__",
    "raise NotImplementedError",
]

[tool.ruff]
target-version = "py312"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "W",
    "I",
    "N",
    "UP",
    "B",
    "A",
    "T20",
    "SIM",
    "TCH",
    "RUF",
]
ignore = [
    "E501",
    "E402",
    "B008",
    "A002",
    "A003",
    "RUF012",
    "F403",
    "F405",
    "B904",
    "N801",
    "TC001",
    "TC003",
    "UP045",
    "UP046",
    "B039",
]

[tool.ruff.lint.per-file-ignores]
"src/app.py" = ["I001"]
"scaffold.py" = ["T201"]
"migrate.py" = ["T201"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
files = ["src", "tests"]
pretty = true
show_error_codes = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
check_untyped_defs = true
no_implicit_optional = true
